backend_api/asset_inventory/requirements.txt
fastapi
uvicorn
python-nmap
httpx
psycopg2-binary